+++ Vec.h
@@ -7,7 +7,7 @@
 
 using std::shared_ptr;
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 class Vec
 {
 public:
@@ -22,104 +22,104 @@ public:
 
     void setAllTo(const T &value);
 
-    T getValue(uint16_t index);
+    T getValue(unsigned int index);
     bool empty() const;
     T norm() const;
     Vec<dim, T> abs() const;
     T getDist(const Vec<dim, T> &vec) const;
     T getSqDist(const Vec<dim, T> &vec) const;
 
-    inline T& operator [] (const uint16_t index) {
+    inline T& operator [] (const unsigned int index) {
         assert (index < dim);
         return m_data[index];
     }
 
-    inline const T& operator [] (const uint16_t index) const {
+    inline const T& operator [] (const unsigned int index) const {
         assert (index < dim);
         return m_data[index];
     }
 
     Vec<dim, T>& operator += (const Vec<dim, T> &vec) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             m_data[i] += vec[i];
         return *this;
     }
 
     Vec<dim, T>& operator += (const T scalar) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             m_data[i] += scalar;
         return *this;
     }
 
     Vec operator + (const Vec<dim, T> &v) const {
         Vec result = Vec<dim, T>::zeros();
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] = (*this)[i] + v[i];
         return result;
     }
 
     Vec& operator -= (const Vec<dim, T> &v) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             (*this)[i] -= v[i];
         return *this;
     }
 
     Vec& operator -= (const T scalar) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             (*this)[i] -= scalar;
         return *this;
     }
 
     Vec operator - (const Vec<dim, T> &v) const {
         Vec result = Vec<dim, T>::zeros();
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] = (*this)[i] - v[i];
         return result;
     }
 
     Vec operator + (const T &s) const {
         Vec<dim, T> result(*this);
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] += s;
         return result;
     }
 
     Vec operator - (const T &s) const {
         Vec<dim, T> result(*this);
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] -= s;
         return result;
     }
 
     Vec& operator *= (const T s) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             (*this)[i] *= s;
         return *this;
     }
 
     Vec operator * (const T &s) const {
         Vec result = Vec<dim, T>::zeros();
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] = (*this)[i] * s;
         return result;
     }
 
     T operator * (const Vec<dim, T> &v) const {
         T result = 0;
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result += (*this)[i] * v[i];
         return result;
     }
 
     Vec& operator /= (const T &s) {
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             (*this)[i] /= s;
         return *this;
     }
 
     Vec operator / (const T &s) const {
         Vec result = Vec<dim, T>::zeros();
-        for (uint16_t i = 0; i < dim; ++i)
+        for (unsigned int i = 0; i < dim; ++i)
             result[i] = (*this)[i] / s;
         return result;
     }
@@ -140,26 +140,26 @@ private:
     T m_data[dim];
 };
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec() {
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         m_data[i] = NAN;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x) {
     assert(dim == 1);
     m_data[0] = x;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x, const T &y) {
     assert(dim == 2);
     m_data[0] = x;
     m_data[1] = y;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x, const T &y, const T &z) {
     assert(dim == 3);
     m_data[0] = x;
@@ -167,7 +167,7 @@ Vec<dim, T>::Vec(const T &x, const T &y, const T &z) {
     m_data[2] = z;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx) {
     assert(dim == 4);
     m_data[0] = x;
@@ -176,7 +176,7 @@ Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx) {
     m_data[3] = rx;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx, const T &ry) {
     assert(dim == 5);
     m_data[0] = x;
@@ -186,7 +186,7 @@ Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx, const T &ry) {
     m_data[4] = ry;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx, const T &ry, const T &rz) {
     assert(dim == 6);
     m_data[0] = x;
@@ -197,56 +197,56 @@ Vec<dim, T>::Vec(const T &x, const T &y, const T &z, const T &rx, const T &ry, c
     m_data[5] = rz;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T>::Vec(const T data[]) {
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         (*this)[i] = data[i];
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 void Vec<dim, T>::setAllTo(const T &value) {
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         (*this)[i] = value;
 }
 
-template<uint16_t dim, typename T>
-T Vec<dim, T>:: getValue(uint16_t index){
+template<unsigned int dim, typename T>
+T Vec<dim, T>:: getValue(unsigned int index){
     return m_data[index];
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 bool Vec<dim, T>::empty() const{
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         if ((*this)[i] == NAN)
             return true;
     return false;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 T Vec<dim, T>::norm() const {
     T norm = 0;
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         norm += (*this)[i] * (*this)[i];
     return sqrtf(norm);
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 Vec<dim, T> Vec<dim, T>::abs() const {
     Vec<dim, T> result;
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         result[i] = fabs((*this)[i]);
     return result;
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 T Vec<dim, T>::getDist(const Vec<dim, T> &vec) const{
     return sqrtf(getSqDist(vec));
 }
 
-template<uint16_t dim, typename T>
+template<unsigned int dim, typename T>
 T Vec<dim, T>::getSqDist(const Vec<dim, T> &vec) const{
     T dist = 0;
-    for (uint16_t i = 0; i < dim; ++i)
+    for (unsigned int i = 0; i < dim; ++i)
         dist += ((*this)[i] - vec[i]) * ((*this)[i] - vec[i]);
     return dist;
 }
